Addition of my latest Assignment
Assignment 4. This assignment introduced the concept of using and accepting inputs, clientside that is. Inputs I've learned vary from radio, to text inputs. I've also included a button dropdown to allow for more space on my nav bar.so it doesn't look as clustered. I have not learned anything about buttons yet, I did however use a template I found on w3schools and briefly styled it to my liking.

Other updates include the formatting of my other pages to create a consistent layout of the nav bar. (As consistent as it can be considering my one website is hosting all my projects!)